REF: reorganize imports and extract BlueLoading
