+++ screen/lnd/LNDViewAdditionalInvoiceInformation.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useRoute, RouteProp } from '@react-navigation/native';
 import { Share, StyleSheet, View } from 'react-native';
-import { BlueLoading, BlueSpacing20, BlueText } from '../../BlueComponents';
+import { BlueSpacing20, BlueText } from '../../BlueComponents';
 import presentAlert from '../../components/Alert';
 import Button from '../../components/Button';
 import CopyTextToClipboard from '../../components/CopyTextToClipboard';
@@ -12,6 +12,7 @@ import loc from '../../loc';
 import { useStorage } from '../../hooks/context/useStorage';
 import { LightningCustodianWallet } from '../../class';
 import { useExtendedNavigation } from '../../hooks/useExtendedNavigation';
+import { BlueLoading } from '../../components/BlueLoading';
 
 type RouteParams = {
   params: {
