+++ screen/settings/SelfTest.tsx
@@ -12,7 +12,7 @@ import * as BlueElectrum from '../../blue_modules/BlueElectrum';
 import * as encryption from '../../blue_modules/encryption';
 import * as fs from '../../blue_modules/fs';
 import ecc from '../../blue_modules/noble_ecc';
-import { BlueLoading, BlueSpacing20, BlueText } from '../../BlueComponents';
+import { BlueSpacing20, BlueText } from '../../BlueComponents';
 import {
   HDAezeedWallet,
   HDSegwitBech32Wallet,
@@ -27,6 +27,7 @@ import Button from '../../components/Button';
 import SaveFileButton from '../../components/SaveFileButton';
 import loc from '../../loc';
 import { CreateTransactionUtxo } from '../../class/wallets/types.ts';
+import { BlueLoading } from '../../components/BlueLoading.tsx';
 
 const bip32 = BIP32Factory(ecc);
